validit

Struct Valid

pub struct Valid<T>
where
    T: Validate,
{ /* private fields */ }

A wrapper of T that validate the state of T every time accessing it.

It validates the state before accessing it, i.e., if when a invalid state is written to it, it wonโt panic until next time accessing it.

The validation is turned on only when debug_assertions is enabled.

An example of defining field a whose value must not exceed 10.

struct Foo { a: u64 }
impl Validate for Foo {
    fn validate(&self) -> Result<(), Box<dyn Error>> {
        less_equal!(self.a, 10);
        Ok(())
    }
}

let f = Valid::new(Foo { a: 20 });
let _x = f.a; // panic: panicked at 'invalid state: expect: self.a(20) <= 10(10) ...

Validation is triggered when:

Deref or DerefMut is called. This is the major use case.

Clone is called.

PartialEq or Eq is called.

PartialOrd or Ord is called.

Hash is called.

Validation is not triggered when:

Copy: Because it is just a byte copy.

Debug and Display: for being able to examine the value for debugging.

Implementations

impl<T: Validate> Valid<T>

pub fn new(inner: T) -> Self

Create a new Valid<T>.

pub fn enable_validation(&mut self, enabled: bool)

Set whether to validate the state when accessing it.

pub fn is_enabled(&self) -> bool

Get whether to validate the state when accessing it.

pub fn as_ref(&self) -> Valid<&T>

Return a reference to the wrapped value: Valid<&T>.

pub fn into_inner(self) -> T

Consume self and return the wrapped value.

This does NOT validate the state.